frontend and csrf cors

+++ spring-security-backend/src/main/java/com/malhar/springsecuritybackend/repository/CustomerRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
 
     List<Customer> findByEmail(String email);
+    
 }
